asm: allow providing a temporary buffer when marshaling an instruction
When marshaling an instruction, we end up creating an instruction size
buffer that we then pass to the writer. The writer being an interface,
the escape analysis we instruct the compiler to allocate the temporary
buffer to the heap.

When marshaling a big slice of instructions, this ends up allocating
a lot of temporary buffers. This commit adds the ability to provide to
reuse the same temporary buffer for each instruction in a slice.

                       │   old.txt   │               new.txt               │
                       │   sec/op    │   sec/op     vs base                │
InstructionsMarshal-10   157.7µ ± 4%   126.5µ ± 0%  -19.79% (p=0.000 n=10)

                       │    old.txt    │              new.txt               │
                       │     B/op      │    B/op     vs base                │
InstructionsMarshal-10   32797.50 ± 0%   32.00 ± 0%  -99.90% (p=0.000 n=10)

                       │    old.txt    │              new.txt               │
                       │   allocs/op   │ allocs/op   vs base                │
InstructionsMarshal-10   4096.000 ± 0%   1.000 ± 0%  -99.98% (p=0.000 n=10)

Signed-off-by: Paul Cacheux <paul.cacheux@datadoghq.com>